requirements.txt
aiohttp==3.7.4.post0
apturl==0.5.2
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bcrypt==3.1.7
bitarray==1.2.2
blinker==1.4
bplib==0.0.6
Brlapi==0.7.0
cached-property==1.5.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.3
cmdtest==0.32+git
coconut-lib==1.3.1
colorama==0.4.3
command-not-found==0.3
coverage==6.4.1
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fasteners==0.14.1
Flask==2.1.2
Flask-PyMongo==2.3.0
fonttools==4.33.3
future==0.18.2
hexbytes==0.2.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonpickle==2.1.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.2
monotonic==1.5
more-itertools==4.2.0
msgpack-python==0.5.6
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.10.4
numpy==1.21.3
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.4.2
paramiko==2.6.0
parsimonious==0.8.1
petlib==0.0.45
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.0.0
protobuf==3.19.0
py==1.11.0
py-ecc==5.2.0
py-solc-x==1.1.1
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.11.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.18.0
pytest==7.1.2
pytest-cov==3.0.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
scipy==1.8.1
seaborn==0.11.2
SecretStorage==2.3.1
semantic-version==2.8.5
simplejson==3.16.0
six==1.14.0
systemd-python==234
texttable==1.6.2
tomli==2.0.1
toolz==0.11.1
ttystatus==0.38
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
varint==1.0.2
wadllib==1.3.3
web3==5.25.0
websocket-client==0.53.0
websockets==9.1
Werkzeug==2.1.2
xkit==0.0.0
yarl==1.7.0
zipp==1.0.0